+++ src/main/java/com/example/gestorancianato/Repositories/CategoriaMedicamentoRepository.java
@@ -4,7 +4,9 @@ import com.example.gestorancianato.Entities.CategoriaMedicamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface CategoriaMedicamentoRepository extends JpaRepository<CategoriaMedicamento, Integer> {
+public interface CategoriaMedicamentoRepository extends JpaRepository<CategoriaMedicamento, Long> {
 
 }
