+++ src/main/java/com/example/gestorancianato/Repositories/SuministroRepository.java
@@ -9,11 +9,11 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface SuministroRepository extends JpaRepository<Suministro, Integer> {
+public interface SuministroRepository extends JpaRepository<Suministro, Long> {
     List<Suministro> findByFechaSuministroBetween(LocalDate startDate, LocalDate endDate);
 
     //Este es para listar los suministros por adulto mayor
-    List<Suministro> findSuministrosByAdultoMayor_Cedula(Integer cedula);
+    List<Suministro> findSuministrosByAdultoMayor_Cedula(Long cedula);
 
     @Query ("SELECT s FROM Suministro s WHERE s.medicamento.nombre = :medicamento")
     List<Suministro> findSuministrosByMedicamento(String medicamento);
