+++ src/main/java/com/example/gestorancianato/Services/CondicionMedicaService.java
@@ -6,11 +6,11 @@ import com.example.gestorancianato.Dtos.CondicionMedicaDto;
 import java.util.List;
 
 public interface CondicionMedicaService {
-    CondicionMedicaDto getCondicionById(Integer id);
+    CondicionMedicaDto getCondicionById(Long id);
     CondicionMedicaDto createCondicion(CondicionMedicaDto condicionMedica);
-    CondicionMedicaDto updateCondicion(Integer id, CondicionMedicaDto condicionMedica);
+    CondicionMedicaDto updateCondicion(Long id, CondicionMedicaDto condicionMedica);
 
-    void deleteCondicionById(Integer id);
+    void deleteCondicionById(Long id);
     List<CondicionMedicaDto> getAllCondiciones();
 
     CondicionMedicaDto getCondicionByNombre(String nombre);
